% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferorg.R
\name{autoconvert}
\alias{autoconvert}
\title{Automatically convert genes to a given format}
\usage{
autoconvert(gene_vec, to = c("symbol", "ensgene", "entrez"))
}
\arguments{
\item{gene_vec}{Vector of input gene IDs}

\item{to}{ID format to convert gene IDs to}
}
\value{
A vector of the same size as gene_vec with the mapped IDs. Any
IDs that are unable to be mapped are converted to \code{NA}.
}
\description{
Infers both the organism and genome ID format before converting
to the desired format
}
\examples{
autoconvert(c("Tapbp", "B2m"), to='ensgene')
autoconvert(c(3105, 3106), to='symbol')
}
