% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferorg.R
\name{inferorg}
\alias{inferorg}
\title{Infer the organism and ID format given a list of genes.}
\usage{
inferorg(gene_vec)
}
\arguments{
\item{gene_vec}{Vector of input gene IDs to guess the organism and format of}
}
\value{
A list with four entries:
\itemize{
\item \code{organism} The inferred organism.
\item \code{format} The inferred ID format.
\item \code{confidence_organism} The confidence in the organism inferred, between 0 and 1.
\item \code{confidence_format} The confidence in the format inferred, between 0 and 1.
}
}
\description{
Infer the organism and ID format given a list of genes.
}
\details{
This function takes in a vector of gene IDs, and guesses the format, which is one of:
\itemize{
\item \code{symbol} (e.g. PTPRC)
\item \code{ensgene} (e.g. ENSG00000081237)
\item \code{entrez} (e.g. 5788)
}
along with the organism, which is one of:
\itemize{
\item \code{human}
\item \code{mouse}
\item \code{fruit_fly}
\item \code{macaque}
\item \code{worm}
\item \code{chicken}
\item \code{rat}
}
In addition, a confidence score for both the organism and format is computed
(see vignette for details).
}
\examples{
genes <- c("HLA-A", "HLA-B", "HLA-C")
inferorg(genes)
}
